Have download link load the raw version of the PDF
Currently the "Here's the download shortcut:" link loads the PDF within the GitHub UI. It does not download the PDF. This is a bit of a pain as the PDF is 3 MB (takes a while to load) and 124 pages (can't navigate that well when the PDF is loaded inline). 

This just points to the raw PDF file, which will download the file as an attachment.