replacing a dead link with a live one (#107)
* Swapped out a missing link

A blog post no longer exists for batch vs. stream processing. I figured a live link is more helpful then a dead one. We can always swap it out for content from Andreas, when such content is created.

* Reformatting

I moved the link down a line. This makes the small paragraph look better.

* Changed Suggested Link to Relevant Link

Instead of sending the reader to a different publisher, the new link will keep them within the Andreas world of content.

* Removing Dead Link

I removed a dead link, as the ultimate document is no longer hosted on Salesforce via Cisco.

I also made a minor adjustment to the next sentence. I believe that my adjustment provides a reason to click on the link.